Restructure ONNX chat server
