+++ apps/README.md
@@ -1,40 +1,51 @@
 # Embedded Servers
 
-The `apps/` directory contains optional HTTP services that expose lightweight APIs used by some nodes.
+The `apps/` directory contains lightweight HTTP services used by the nodes. Each service is self contained so you can run them independently or inside Docker.
 
 ## ONNX Chat Completion Server
 
-`onnx_server.py` starts a minimal OpenAI compatible endpoint. It can run a toy ONNX model or fall back to very simple rules if no model is provided. Use it when you need a local endpoint for the query nodes.
+The `onnx_chat/` folder exposes a tiny OpenAI compatible endpoint. It can load a toy ONNX model if `ONNX_MODEL_PATH` is set, otherwise a simple rule based reply is returned.
 
-Run it with:
+Start the server locally with:
 
 ```bash
-python -m apps.onnx_server
+python -m apps.onnx_chat.main
 ```
 
-Set `ONNX_MODEL_PATH` to point at an ONNX model file if you have one. The server listens on port `8000` by default.
+Or via the convenience entrypoint:
+
+```bash
+comfyai-onnx-server
+```
+
+By default the server listens on port `8000`. Supply `PORT` to override. A `Dockerfile` is also provided so you can build an image with:
+
+```bash
+docker build -t onnx-chat apps/onnx_chat
+```
+
+Run it using:
+
+```bash
+docker run -p 8000:8000 onnx-chat
+```
 
 ## Face Comparison API
 
-`face_api/` hosts a small FastAPI application that compares two face images. It relies on InsightFace and ONNX Runtime to extract embeddings.
+`face_api/` hosts a FastAPI application to compare two face images. It uses InsightFace and ONNX Runtime to extract embeddings.
 
-### Installation
-
-Install the dependencies if you plan to run the API outside Docker:
+Install the dependencies if running outside Docker:
 
 ```bash
-pip install fastapi uvicorn insightface onnxruntime[-gpu] python-multipart
+pip install -r apps/face_api/requirements.txt
 ```
 
-Launch it using:
+Launch it with uvicorn:
 
 ```bash
 uvicorn apps.face_api.main:app --host 0.0.0.0 --port 7860
 ```
 
-Several environment variables let you tune which provider or model is used:
+A `Dockerfile` is provided in the same directory. Build it with `docker build -t face-api apps/face_api` and run with `docker run -p 7860:7860 face-api`.
 
-- `FACE_MODEL_PROVIDERS` – comma separated list of ONNX providers (default: `CUDAExecutionProvider,CPUExecutionProvider`)
-- `FACE_MODEL_NAME` – name of the InsightFace model (default: `buffalo_l`)
-
-The `CompareFacesNode` sends images to this API and receives a similarity score.
+Environment variables allow you to select models or execution providers. Refer to `apps/face_api/main.py` for details.
