+++ apps/onnx_chat/requirements.txt
@@ -1,0 +1,3 @@
+fastapi
+uvicorn
+onnxruntime
